Update personalcar.rb
In this program we can define two types of attributes. Personal name and surname attributes are cannot be change so we define it in as 'reader' way but age and degree situations we define in 'accessor' way because they changes after monthly or yealry. Same thing in car properties. Brand and car number do not but color and km attributes do.